Update day-1 to be able to handle text-based integers
Previously it only looked for numbers (as chars) 1-9. Since the
example-text also includes their text representation (one, two, three,
etc.) I had to make some changes to the code. Now there is a map with
the string value and their corresponding integer. The values are now
stored in a map, and what is returned is instead the first element and
last element of a map, rather than a deque.