Day 2 part two complete
Changed some things to make part 2 work along with part 1. Now part 1
creates all games of a row, where it previously cancelled immediately
upon finding one was impossible of the set. It then compares all games
and if all do not succeed, it returns 0. Rather than a for loop, this is
now done using std::all_of in order for me to practice with it.
Additionally, the sum in part one and part two are now done using an
accumulator function for the same reason.

Part two creates a map to keep track of the three largest orders of
colors, and then returns the sum of all of these.